[FEATURE] Proper page settings inheritance
This adds inheritance of page variables for sub pages using either an inherited page template or the same template as the parent page. In addition a new approach to trim inherited variables has been implemented, greatly improving the predictability - changing a variable in a parent page makes that variable effective on any page down the rootline until either the page template is changed or the variable in question has been intentionally overridden by a page.