[TASK] Retrieve "stored variable" cache by stricter key
This takes into consideration which variable is requested (in any Flux extension, this would be "storage" so no impact there), the paths (these could change dynamically in a ConfigurationProvider on the same page, i.e. in multiple content elements using the same ConfigurationProvider and should prevent some undesired reuse) and finally the section name that contains the variable to retrieve (again in almost all cases would be "Configuration" but also could be dynamically set by ConfigurationProviders).