[TASK] Catch "file not found" Exception and send NOTICE message
This allows the "file not found" Exception to be handled gracefully in the backend - which may happen for sys folders (or other folders which are outside the page tree for whatever reason).

The previously fatal error thrown by Flux itself is now caught by FluidPages (when trying to read the backend layout from the template). Replacing the fatal error with a NOTICE allows it to be hidden from view when Flux debug mode is set to staging or production mode (0 or 1).