[BUGFIX] Use original record as emulated parent for versioned records
This makes inheritance resolving in PageService include the original record as an emulated first parent record in the root line of a record, if said record is a versioned record. The effect propagates all the way through the root line, resolving the original record's parent values for each record in the root line.