[FEATURE] Replace deprecated configuration translation
Since TYPO3 7.6 the previous handling of translated
page configuration has been deprecated and moved
into the compatibility6 extension. To replace this
functionality this commit introduces the flexform fields
 for configuration to the  pages_language_overlay
table and takes care of the overlaying of the affected
configurations. This functionality is disable by default
and can be activated through a toggle in the extension
configuration.

related deprecation: #70138 - Flex form language handling