[TASK] Improve Performance of PageProvider
we should check for the existence of the Cache key
instead of it being empty, because the Inherited Configuration
often tends to be empty, why means, we retry this countless
times without any good reason. This results in a exponential
growth old load time in relation to the amount of fields in a sheet.