[BUGFIX] Fix unchanged inherited properties overriding in sub-pages
The first param passed to the function `getInheritedPropertyValueByDottedPath()`
was a full record array, when it should be a configuration array.

This results in the inherited configuration value always being equal to `null`,
and forcing the value to be submitted in each and every sub-page FlexForm. It
means that the inheritance was break for every page that was updated.

When changing the value of a configuration value for a root page, all its
sub-pages that were saved once were not updated.

This commit fixes this bug by actually sending the configuration to the
function.

A nice add-on would be to create an extension upgrade script that processes the
entire page tree to fix the corrupted data (proposed by @NamelessCoder). :+1: