[TASK] Introduce memoization in `PageService` for FE environment
As it has been discussed in #340, pages handled by FluidPages and containing
dozens of configuration values currently lead to dozens of massive database
queries (like `SELECT * FROM pages WHERE uid='xxx';`), resulting in a huge slow
of the application for the first frontend request (next ones are cached, so they
are faster).

This commit introduces a memoizer in the `PageService`: if the request is in a
frontend environment, it will be used to execute the select-query only once per
page-uid.

Please note that the `RecordService` is still used directly in other classes
that are used only in backend environment, for instance in:
`\FluidTYPO3\Fluidpages\Backend\BackendLayout::postProcessBackendLayout()`.