[FEATURE] Use "static" flag from form for persistent DS caching
The feature works by completely avoiding template parsing
to retrieve a Form, when the Form is marked static. Instead
the form gets saved in a persistent cache and gets associated
with the record (uid and pid) and field name it serves.

This allows the DS to depend on global variables such as
extension configuration and (global) TS, but not other fields
in the page record.

When a template is not marked static it gets added to the
runtime cache for a similar benefit, in particular relevant when
fluidpages parses multiple nodes on large page trees.