[BUGFIX] Avoid root page query in getPageTemplateConfiguration
This fix aborts the loop in PageService->getPageTemplateConfiguration
early if the uid to fetch from the database becomes zero.
Database uids are guaranteed to start from one because uid 0 is reserved
for the virtual page root.

Also added an early return if the initial page record cannot be found
and fixed failing tests due to wrong or incomplete mocked records.