[FEATURE] Hide entire "flex" field if empty with Flux 9.3+
Adds a condition to check if a certain feature class of Flux
exists and if it does, adds this as displayCond. The condition
will only allow the field to be shown if there are actually
fields inside the DS.

Note that the feature being used is only available on Flux
9.3 (which at this time is not released yet). The minimum
requirement for Flux version is however not increased -
the friendlier solution of just not using the feature if not
available is implemented.

Close: #374