[BUGFIX] Default to "grid" layout name
When the record is invalid or contains no grid, but
Fluidpages is selected as layout, return an empty
grid with the name "grid".

This avoids a reset of the "backend_layout" fields
in the page record when the grid names do not
match the previous value. By default, all grids in
Flux are named "grid" unless explicitly renamed -
so we choose "grid" as fallback name.

Close: #411